pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autopep8>=2.3.2",
    "fastapi>=0.133.0",
    "inngest>=0.5.15",
    "langchain>=1.2.12",
    "langchain-ollama>=1.0.1",
    "litellm>=1.82.0",
    "llama-index-core>=0.14.15",
    "llama-index-readers-file>=0.5.6",
    "pydantic>=2.12.5",
    "pyproject-toml>=0.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
